Allow to use floats in damage pattern editor

+++ eos/db/saveddata/damagePattern.py
@@ -17,7 +17,7 @@
 # along with eos.  If not, see <http://www.gnu.org/licenses/>.
 # ===============================================================================
 
-from sqlalchemy import Table, Column, Integer, ForeignKey, String, DateTime
+from sqlalchemy import Table, Column, Integer, Float, ForeignKey, String, DateTime
 from sqlalchemy.orm import mapper
 import datetime
 
@@ -27,10 +27,10 @@ from eos.saveddata.damagePattern import DamagePattern
 damagePatterns_table = Table("damagePatterns", saveddata_meta,
                              Column("ID", Integer, primary_key=True),
                              Column("name", String),
-                             Column("emAmount", Integer),
-                             Column("thermalAmount", Integer),
-                             Column("kineticAmount", Integer),
-                             Column("explosiveAmount", Integer),
+                             Column("emAmount", Float),
+                             Column("thermalAmount", Float),
+                             Column("kineticAmount", Float),
+                             Column("explosiveAmount", Float),
                              Column("ownerID", ForeignKey("users.ID"), nullable=True),
                              Column("created", DateTime, nullable=True, default=datetime.datetime.now),
                              Column("modified", DateTime, nullable=True, onupdate=datetime.datetime.now)
